+++ HISTORY.rst
@@ -10,5 +10,12 @@ History
 
 0.2.0 (near future)
 ---------------------
+
+* new StereoNet class for Schmidt projection
+* mplstereonet dependency depreceated
+
 * new Pair and Fault classes to manipulate attached data
 * new uniform_lin and uniform_fol Group methods
+* abs for Group implemented to calculate euclidean norms
+* normalized method for group
+